# Add fast path for reading consecutive strings/byte arrays