--- a/src/MessagePack.UniversalCodeGenerator/Generator/FormatterTemplate.cs
+++ b/src/MessagePack.UniversalCodeGenerator/Generator/FormatterTemplate.cs
@@ -374,7 +374,7 @@ if(objInfo.HasIMessagePackSerializationCallbackReceiver && objInfo.NeedsCastOnBe
             
             #line default
             #line hidden
-            this.Write(@"                ReadOnlySequence<byte> stringKey = reader.ReadStringSequence().Value;
+            this.Write(@"                ReadOnlySpan<byte> stringKey = Internal.CodeGenHelpers.ReadStringSpan(ref reader);
                 int key;
                 if (!this.____keyMapping.TryGetValue(stringKey, out key))
                 {