--- a/src/MessagePack.UniversalCodeGenerator/Generator/FormatterTemplate.tt
+++ b/src/MessagePack.UniversalCodeGenerator/Generator/FormatterTemplate.tt
@@ -104,7 +104,7 @@ namespace <#= Namespace #>
             for (int i = 0; i < length; i++)
             {
 <# if(objInfo.IsStringKey) { #>
-                ReadOnlySequence<byte> stringKey = reader.ReadStringSegment().Value;
+                ReadOnlySequence<byte> stringKey = reader.ReadStringSequence().Value;
                 int key;
                 if (!this.____keyMapping.TryGetValue(stringKey, out key))
                 {