--- a/src/MessagePack.UniversalCodeGenerator/publish.bat
+++ b/src/MessagePack.UniversalCodeGenerator/publish.bat
@@ -1,3 +1,3 @@
-dotnet publish MessagePack.UniversalCodeGenerator.csproj -c Release --self-contained -r win-x64 -o ./bin/MessagePackUniversalCodeGenerator/win-x64
-dotnet publish MessagePack.UniversalCodeGenerator.csproj -c Release --self-contained -r linux-x64 -o ./bin/MessagePackUniversalCodeGenerator/linux-x64
-dotnet publish MessagePack.UniversalCodeGenerator.csproj -c Release --self-contained -r osx-x64 -o ./bin/MessagePackUniversalCodeGenerator/osx-x64
+dotnet publish %~dp0MessagePack.UniversalCodeGenerator.csproj -c Release --self-contained -r win-x64
+dotnet publish %~dp0MessagePack.UniversalCodeGenerator.csproj -c Release --self-contained -r linux-x64
+dotnet publish %~dp0MessagePack.UniversalCodeGenerator.csproj -c Release --self-contained -r osx-x64