This does 3 things:

1. Rename `MessagePackReader.ReadStringSegment` to `ReadStringSequence`
1. Add `MessagePackReader.TryReadStringSpan`
1. Speed up string key deserialization by avoiding `ReadOnlySequence<byte>.Slice` in fast path

**Before**

|             Method |       Mean |
|------------------- |-----------:|
|             IntKey |   230.8 ns |
|       OldStringKey |   200.5 ns |
|           Automata | 2,233.2 ns |
|          Hashtable | 1,551.9 ns |
| AutomataInlineEmit | 1,724.8 ns |
**After**

|             Method |       Mean |
|------------------- |-----------:|
|             IntKey |   216.4 ns |
|       OldStringKey |   150.1 ns |
|           Automata | 1,705.1 ns |
|          Hashtable | 1,444.3 ns |
| AutomataInlineEmit |   498.2 ns |

Clearly the biggest improvement is to the `AutomataInlineEmit` scenario.

Fixes #513