@neuecc 2019-11-05
thanks, nice improvement!

Can you modify codegen uses it?
`ReadOnlySequence<byte> stringKey = reader.ReadStringSequence().Value;` to, equivant of
`var stringKey = `CodeGenHelpers.ReadStringSpan(ref reader)`