@AArnott 2019-11-05
@neuecc I updated the code generators, but only one of the two generated files that are checked into the source tree. I'll update this when #618 is addressed.